backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
click==8.3.1
colorama==0.4.6
ecdsa==0.19.1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.135.1
google-genai==1.49.0
greenlet==3.3.2
h11==0.16.0
httptools==0.7.1
idna==3.11
lxml==6.0.2
openpyxl==3.1.5
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.2
PyPDF2==3.0.1
pydantic==2.12.5
pydantic_core==2.41.5
python-docx==1.2.0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.47
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
watchfiles==1.1.1
websockets==15.0.1
